% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xarf.R
\name{inferType}
\alias{inferType}
\title{Infer type of XARF based on vector \code{x}}
\usage{
inferType(x)
}
\arguments{
\item{x}{is a vector that type to be determined}
}
\value{

}
\description{
Infer type of XARF based on vector \code{x}
}
\examples{
inferType(1:10)
}
