% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xarf.R
\name{write.xarf}
\alias{write.xarf}
\title{Data Output to XARF}
\usage{
write.xarf(x, dbID, file = "", dbDescription = NULL,
  attrDescription = NULL)
}
\arguments{
\item{x}{the data.frame to be written}

\item{dbID}{a unique identifier of the dataset}

\item{file}{either a character string naming a file or a connection open for writing. "" indicates output to the console.}

\item{dbDescription}{a list specifying the description of the dataset}

\item{attrDescription}{a vector of descriptions for every attribute}
}
\description{
Function saves a data.frame like objct to XARF format.
More details on format can be found in \url{https://bitbucket.org/realKD/realkd/wiki/model/data/xarf}.
}
\details{
The function takes the dataset and writes it to file \code{file}.
The names of the rows are written as the name attribute.
The attributes are written in the order given in \code{attrDescription}.
If an attribute is missing from \code{attrDescription}, it is written at the end of the dataset.
}
\examples{
write.xarf(mtcars, file="",
  "mtcars",
  dbDescription=list(
    title="Motor Trend Car Road Tests",
    version="1.0.0",
    source="Basic R dataset 'mtcars'"),
  attrDescription=list(
    list(id="mpg",name="Miles/(US) gallon",type="real",description="The distance in miles a car can go with a single gallon of petrolium"),
    list(id="cyl",name="Number of cylinders",type="integer"),
    list(id="disp",name="Displacement (cu.in.)")
  ))
}
